+++ Kernel/Modules/AdminTemplate.pm
@@ -103,6 +103,16 @@ sub Run {
             }
         }
 
+        # check if a standard template exist with this name
+        my $NameExists
+            = $Self->{StandardTemplateObject}
+            ->NameExistsCheck( Name => $GetParam{Name}, ID => $GetParam{ID} );
+
+        if ($NameExists) {
+            $Errors{NameExists} = 1;
+            $Errors{'NameInvalid'} = 'ServerError';
+        }
+
         # if no errors occurred
         if ( !%Errors ) {
 
@@ -214,6 +224,14 @@ sub Run {
             if ( !$GetParam{$Needed} ) {
                 $Errors{ $Needed . 'Invalid' } = 'ServerError';
             }
+        }
+
+        # check if a standard template exists with this name
+        my $NameExists
+            = $Self->{StandardTemplateObject}->NameExistsCheck( Name => $GetParam{Name} );
+        if ($NameExists) {
+            $Errors{NameExists} = 1;
+            $Errors{'NameInvalid'} = 'ServerError';
         }
 
         # if no errors occurred
@@ -369,6 +387,14 @@ sub _Edit {
     }
     else {
         $Self->{LayoutObject}->Block( Name => 'HeaderAdd' );
+    }
+
+    # show appropriate messages for ServerError
+    if ( defined $Param{Errors}->{NameExists} && $Param{Errors}->{NameExists} == 1 ) {
+        $Self->{LayoutObject}->Block( Name => 'ExistNameServerError' );
+    }
+    else {
+        $Self->{LayoutObject}->Block( Name => 'NameServerError' );
     }
 
     # add rich text editor